Extract ProtocolsHandlerSelect
