+++ core/src/protocols_handler/mod.rs
@@ -19,27 +19,29 @@
 // DEALINGS IN THE SOFTWARE.
 
 use crate::{
-    either::EitherOutput,
     nodes::handled_node::{NodeHandler, NodeHandlerEndpoint, NodeHandlerEvent},
     upgrade::{
         self,
         InboundUpgrade,
-        InboundUpgradeExt,
         OutboundUpgrade,
         InboundUpgradeApply,
         OutboundUpgradeApply,
-        EitherUpgrade,
-        OrUpgrade
     }
 };
 use futures::prelude::*;
-use std::{io, marker::PhantomData, time::Duration};
+use std::{io, time::Duration};
 use tokio_io::{AsyncRead, AsyncWrite};
 use tokio_timer::Timeout;
 
 pub use self::dummy::DummyProtocolsHandler;
+pub use self::map_in::MapInEvent;
+pub use self::map_out::MapOutEvent;
+pub use self::select::ProtocolsHandlerSelect;
 
 mod dummy;
+mod map_in;
+mod map_out;
+mod select;
 
 /// Handler for a set of protocols for a specific connection with a remote.
 ///
@@ -155,11 +157,7 @@ pub trait ProtocolsHandler {
         Self: Sized,
         TMap: Fn(&TNewIn) -> Option<&Self::InEvent>,
     {
-        MapInEvent {
-            inner: self,
-            map,
-            marker: PhantomData,
-        }
+        MapInEvent::new(self, map)
     }
 
     /// Adds a closure that turns the output event into something else.
@@ -169,7 +167,7 @@ pub trait ProtocolsHandler {
         Self: Sized,
         TMap: FnMut(Self::OutEvent) -> TNewOut,
     {
-        MapOutEvent { inner: self, map }
+        MapOutEvent::new(self, map)
     }
 
     /// Builds an implementation of `ProtocolsHandler` that handles both this protocol and the
@@ -179,10 +177,7 @@ pub trait ProtocolsHandler {
     where
         Self: Sized,
     {
-        ProtocolsHandlerSelect {
-            proto1: self,
-            proto2: other,
-        }
+        ProtocolsHandlerSelect::new(self, other)
     }
 
     /// Creates a builder that will allow creating a `NodeHandler` that handles this protocol
@@ -285,158 +280,6 @@ impl<TConnectionUpgrade, TOutboundOpenInfo, TCustom>
             }
             ProtocolsHandlerEvent::Custom(val) => ProtocolsHandlerEvent::Custom(map(val)),
         }
-    }
-}
-
-/// Wrapper around a protocol handler that turns the input event into something else.
-pub struct MapInEvent<TProtoHandler, TNewIn, TMap> {
-    inner: TProtoHandler,
-    map: TMap,
-    marker: PhantomData<TNewIn>,
-}
-
-impl<TProtoHandler, TMap, TNewIn> ProtocolsHandler for MapInEvent<TProtoHandler, TNewIn, TMap>
-where
-    TProtoHandler: ProtocolsHandler,
-    TMap: Fn(TNewIn) -> Option<TProtoHandler::InEvent>,
-{
-    type InEvent = TNewIn;
-    type OutEvent = TProtoHandler::OutEvent;
-    type Substream = TProtoHandler::Substream;
-    type InboundProtocol = TProtoHandler::InboundProtocol;
-    type OutboundProtocol = TProtoHandler::OutboundProtocol;
-    type OutboundOpenInfo = TProtoHandler::OutboundOpenInfo;
-
-    #[inline]
-    fn listen_protocol(&self) -> Self::InboundProtocol {
-        self.inner.listen_protocol()
-    }
-
-    #[inline]
-    fn inject_fully_negotiated_inbound(
-        &mut self,
-        protocol: <Self::InboundProtocol as InboundUpgrade<Self::Substream>>::Output
-    ) {
-        self.inner.inject_fully_negotiated_inbound(protocol)
-    }
-
-    #[inline]
-    fn inject_fully_negotiated_outbound(
-        &mut self,
-        protocol: <Self::OutboundProtocol as OutboundUpgrade<Self::Substream>>::Output,
-        info: Self::OutboundOpenInfo
-    ) {
-        self.inner.inject_fully_negotiated_outbound(protocol, info)
-    }
-
-    #[inline]
-    fn inject_event(&mut self, event: TNewIn) {
-        if let Some(event) = (self.map)(event) {
-            self.inner.inject_event(event);
-        }
-    }
-
-    #[inline]
-    fn inject_dial_upgrade_error(&mut self, info: Self::OutboundOpenInfo, error: io::Error) {
-        self.inner.inject_dial_upgrade_error(info, error)
-    }
-
-    #[inline]
-    fn inject_inbound_closed(&mut self) {
-        self.inner.inject_inbound_closed()
-    }
-
-    #[inline]
-    fn shutdown(&mut self) {
-        self.inner.shutdown()
-    }
-
-    #[inline]
-    fn poll(
-        &mut self,
-    ) -> Poll<
-        Option<ProtocolsHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::OutEvent>>,
-        io::Error,
-    > {
-        self.inner.poll()
-    }
-}
-
-/// Wrapper around a protocol handler that turns the output event into something else.
-pub struct MapOutEvent<TProtoHandler, TMap> {
-    inner: TProtoHandler,
-    map: TMap,
-}
-
-impl<TProtoHandler, TMap, TNewOut> ProtocolsHandler for MapOutEvent<TProtoHandler, TMap>
-where
-    TProtoHandler: ProtocolsHandler,
-    TMap: FnMut(TProtoHandler::OutEvent) -> TNewOut,
-{
-    type InEvent = TProtoHandler::InEvent;
-    type OutEvent = TNewOut;
-    type Substream = TProtoHandler::Substream;
-    type InboundProtocol = TProtoHandler::InboundProtocol;
-    type OutboundProtocol = TProtoHandler::OutboundProtocol;
-    type OutboundOpenInfo = TProtoHandler::OutboundOpenInfo;
-
-    #[inline]
-    fn listen_protocol(&self) -> Self::InboundProtocol {
-        self.inner.listen_protocol()
-    }
-
-    #[inline]
-    fn inject_fully_negotiated_inbound(
-        &mut self,
-        protocol: <Self::InboundProtocol as InboundUpgrade<Self::Substream>>::Output
-    ) {
-        self.inner.inject_fully_negotiated_inbound(protocol)
-    }
-
-    #[inline]
-    fn inject_fully_negotiated_outbound(
-        &mut self,
-        protocol: <Self::OutboundProtocol as OutboundUpgrade<Self::Substream>>::Output,
-        info: Self::OutboundOpenInfo
-    ) {
-        self.inner.inject_fully_negotiated_outbound(protocol, info)
-    }
-
-    #[inline]
-    fn inject_event(&mut self, event: Self::InEvent) {
-        self.inner.inject_event(event)
-    }
-
-    #[inline]
-    fn inject_dial_upgrade_error(&mut self, info: Self::OutboundOpenInfo, error: io::Error) {
-        self.inner.inject_dial_upgrade_error(info, error)
-    }
-
-    #[inline]
-    fn inject_inbound_closed(&mut self) {
-        self.inner.inject_inbound_closed()
-    }
-
-    #[inline]
-    fn shutdown(&mut self) {
-        self.inner.shutdown()
-    }
-
-    #[inline]
-    fn poll(
-        &mut self,
-    ) -> Poll<
-        Option<ProtocolsHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::OutEvent>>,
-        io::Error,
-    > {
-        Ok(self.inner.poll()?.map(|ev| {
-            ev.map(|ev| match ev {
-                ProtocolsHandlerEvent::Custom(ev) => ProtocolsHandlerEvent::Custom((self.map)(ev)),
-                ProtocolsHandlerEvent::OutboundSubstreamRequest { upgrade, info } => {
-                    ProtocolsHandlerEvent::OutboundSubstreamRequest { upgrade, info }
-                }
-            })
-        }))
     }
 }
 
@@ -647,121 +490,6 @@ where
             }
             Async::Ready(None) => return Ok(Async::Ready(None)),
             Async::NotReady => (),
-        };
-
-        Ok(Async::NotReady)
-    }
-}
-
-/// Implementation of `ProtocolsHandler` that combines two protocols into one.
-#[derive(Debug, Clone)]
-pub struct ProtocolsHandlerSelect<TProto1, TProto2> {
-    proto1: TProto1,
-    proto2: TProto2,
-}
-
-impl<TSubstream, TProto1, TProto2, TProto1Out, TProto2Out>
-    ProtocolsHandler for ProtocolsHandlerSelect<TProto1, TProto2>
-where
-    TProto1: ProtocolsHandler<Substream = TSubstream>,
-    TProto2: ProtocolsHandler<Substream = TSubstream>,
-    TSubstream: AsyncRead + AsyncWrite,
-    TProto1::InboundProtocol: InboundUpgrade<TSubstream, Output = TProto1Out>,
-    TProto2::InboundProtocol: InboundUpgrade<TSubstream, Output = TProto2Out>,
-    TProto1::OutboundProtocol: OutboundUpgrade<TSubstream, Output = TProto1Out>,
-    TProto2::OutboundProtocol: OutboundUpgrade<TSubstream, Output = TProto2Out>
-{
-    type InEvent = EitherOutput<TProto1::InEvent, TProto2::InEvent>;
-    type OutEvent = EitherOutput<TProto1::OutEvent, TProto2::OutEvent>;
-    type Substream = TSubstream;
-    type InboundProtocol = OrUpgrade<TProto1::InboundProtocol, TProto2::InboundProtocol>;
-    type OutboundProtocol = EitherUpgrade<TProto1::OutboundProtocol, TProto2::OutboundProtocol>;
-    type OutboundOpenInfo = EitherOutput<TProto1::OutboundOpenInfo, TProto2::OutboundOpenInfo>;
-
-    #[inline]
-    fn listen_protocol(&self) -> Self::InboundProtocol {
-        let proto1 = self.proto1.listen_protocol();
-        let proto2 = self.proto2.listen_protocol();
-        proto1.or_inbound(proto2)
-    }
-
-    fn inject_fully_negotiated_outbound(&mut self, protocol: <Self::OutboundProtocol as OutboundUpgrade<TSubstream>>::Output, endpoint: Self::OutboundOpenInfo) {
-        match (protocol, endpoint) {
-            (EitherOutput::First(protocol), EitherOutput::First(info)) =>
-                self.proto1.inject_fully_negotiated_outbound(protocol, info),
-            (EitherOutput::Second(protocol), EitherOutput::Second(info)) =>
-                self.proto2.inject_fully_negotiated_outbound(protocol, info),
-            (EitherOutput::First(_), EitherOutput::Second(_)) =>
-                panic!("wrong API usage: the protocol doesn't match the upgrade info"),
-            (EitherOutput::Second(_), EitherOutput::First(_)) =>
-                panic!("wrong API usage: the protocol doesn't match the upgrade info")
-        }
-    }
-
-    fn inject_fully_negotiated_inbound(&mut self, protocol: <Self::InboundProtocol as InboundUpgrade<TSubstream>>::Output) {
-        match protocol {
-            EitherOutput::First(protocol) =>
-                self.proto1.inject_fully_negotiated_inbound(protocol),
-            EitherOutput::Second(protocol) =>
-                self.proto2.inject_fully_negotiated_inbound(protocol)
-        }
-    }
-
-    #[inline]
-    fn inject_event(&mut self, event: Self::InEvent) {
-        match event {
-            EitherOutput::First(event) => self.proto1.inject_event(event),
-            EitherOutput::Second(event) => self.proto2.inject_event(event),
-        }
-    }
-
-    #[inline]
-    fn inject_inbound_closed(&mut self) {
-        self.proto1.inject_inbound_closed();
-        self.proto2.inject_inbound_closed();
-    }
-
-    #[inline]
-    fn inject_dial_upgrade_error(&mut self, info: Self::OutboundOpenInfo, error: io::Error) {
-        match info {
-            EitherOutput::First(info) => self.proto1.inject_dial_upgrade_error(info, error),
-            EitherOutput::Second(info) => self.proto2.inject_dial_upgrade_error(info, error),
-        }
-    }
-
-    #[inline]
-    fn shutdown(&mut self) {
-        self.proto1.shutdown();
-        self.proto2.shutdown();
-    }
-
-    fn poll(&mut self) -> Poll<Option<ProtocolsHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::OutEvent>>, io::Error> {
-        match self.proto1.poll()? {
-            Async::Ready(Some(ProtocolsHandlerEvent::Custom(event))) => {
-                return Ok(Async::Ready(Some(ProtocolsHandlerEvent::Custom(EitherOutput::First(event)))));
-            },
-            Async::Ready(Some(ProtocolsHandlerEvent::OutboundSubstreamRequest { upgrade, info})) => {
-                return Ok(Async::Ready(Some(ProtocolsHandlerEvent::OutboundSubstreamRequest {
-                    upgrade: EitherUpgrade::A(upgrade),
-                    info: EitherOutput::First(info),
-                })));
-            },
-            Async::Ready(None) => return Ok(Async::Ready(None)),
-            Async::NotReady => ()
-        };
-
-        match self.proto2.poll()? {
-            Async::Ready(Some(ProtocolsHandlerEvent::Custom(event))) => {
-                return Ok(Async::Ready(Some(ProtocolsHandlerEvent::Custom(EitherOutput::Second(event)))));
-            },
-            Async::Ready(Some(ProtocolsHandlerEvent::OutboundSubstreamRequest { upgrade, info })) => {
-                return Ok(Async::Ready(Some(ProtocolsHandlerEvent::OutboundSubstreamRequest {
-                    upgrade: EitherUpgrade::B(upgrade),
-                    info: EitherOutput::Second(info),
-                })));
-            },
-            Async::Ready(None) => return Ok(Async::Ready(None)),
-            Async::NotReady => ()
         };
 
         Ok(Async::NotReady)
